backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.8"
dependencies = [
    "alembic>=1.14.1",
    "argon2-cffi>=23.1.0",
    "asyncpg>=0.30.0",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.115.8",
    "jose>=1.0.0",
    "pydantic>=2.10.6",
    "pyjwt>=2.10.1",
    "redis>=5.2.1",
    "sqlalchemy[asyncio]>=2.0.38",
    "websocket>=0.2.1",
]
